pyproject.toml
[project]
name = "pixeltovoxelprojector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "astropy>=7.1.0",
    "gradio>=5.38.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pillow>=11.3.0",
    "pyvista>=0.45.3",
    "torch>=2.7.1",
]
